bug fix for non-functional spm_filter()
spm_filter assigned fdata to Y, modified Y, and then returned fdata, suggesting Y was meant to modify fdata in place, but Y was a copy, not a view of fdata, so fdata was never updated at all and spm_filter simply returned its input numodified. The fix implements an inplace updating of fdata instead.